[feat] 뷰가 켜지는 모션 제거
